+++ src/UiPath.Caching.Azure/GlobalUsings.cs
@@ -5,7 +5,6 @@ global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.DependencyInjection.Extensions;
 global using Microsoft.Extensions.Options;
-global using StackExchange.Redis;
 global using UiPath.Caching.Config;
 global using UiPath.Caching.Redis;