+++ src/UiPath.Caching.Queue/GlobalUsings.cs
@@ -5,7 +5,6 @@ global using Microsoft.Extensions.DependencyInjection.Extensions;
 global using Microsoft.Extensions.Logging;
 global using Microsoft.Extensions.Logging.Abstractions;
 global using Microsoft.Extensions.Options;
-global using StackExchange.Redis;
 global using UiPath.Caching.Policies;
 global using UiPath.Caching.Redis;
 global using UiPath.Caching.Telemetry;