+++ src/UiPath.Caching.Queue/RedisQueueCacheProvider.cs
@@ -10,7 +10,7 @@ public sealed class RedisQueueCacheProvider : IQueueCacheProvider
     private readonly CacheOptions _cacheOptions;
     private readonly RedisSetCacheOptions _setCacheOptions;
     private readonly IRedisConnector _redis;
-    private readonly ISerializerProxy<RedisValue> _serializerProxy;
+    private readonly ISerializerProxy<byte[]> _serializerProxy;
     private readonly IResiliencePipelineProvider _resiliencePipelineProvider;
     private readonly ICachingTelemetryProvider _cachingTelemetryProvider;
     private readonly ILoggerFactory _loggerFactory;
@@ -26,7 +26,7 @@ public sealed class RedisQueueCacheProvider : IQueueCacheProvider
         IOptions<CacheOptions> cacheOptions,
         IOptions<RedisSetCacheOptions> setCacheOptions,
         IRedisConnector redis,
-        ISerializerProxy<RedisValue> serializerProxy,
+        ISerializerProxy<byte[]> serializerProxy,
         IResiliencePipelineProvider resiliencePipelineProvider,
         ICachingTelemetryProvider cachingTelemetryProvider,
         ILoggerFactory loggerFactory,